requirements.txt
flask
Authlib
requests
python-dotenv